requirements.txt
discord.py
python-dotenv
requests
tzdata
black